Fix a situation where the new tab took longer to load than the time to switch to the one after it (i.e. next tab loads in 5 seconds, but the tab after that should appear in 3 seconds).
In this case, by the time the tab after the next one appears, there are no
"left tabs" to remove. No movement of the "old tab" is seen, and it is not
removed from HTML, causing many tabs to stack up. The fix is to only start the
timer for loading the next tab when the current one is done loading and
animating, guaranteeing that by the time the next tab is supposed to load,
there is exactly one tab and that one is marked "left".